Conflicts:
	Documentation/devicetree/bindings/power/renesas,sysc-rcar.txt
	arch/arm/boot/dts/r8a7779.dtsi
	arch/arm/boot/dts/r8a7790-lager.dts
	arch/arm/boot/dts/r8a7790.dtsi
	arch/arm/boot/dts/r8a7791-porter.dts
	arch/arm/boot/dts/r8a7791.dtsi
	arch/arm/boot/dts/r8a7793.dtsi
	arch/arm/boot/dts/r8a7794.dtsi
	arch/arm/mach-shmobile/Kconfig
	arch/arm/mach-shmobile/pm-r8a7779.c
	arch/arm/mach-shmobile/pm-rcar-gen2.c
	arch/arm/mach-shmobile/pm-rcar.h
	arch/arm/mach-shmobile/smp-r8a7779.c
	arch/arm/mach-shmobile/smp-r8a7790.c
	arch/arm64/boot/dts/renesas/r8a7795-salvator-x.dts
	arch/arm64/boot/dts/renesas/r8a7795.dtsi
	drivers/clk/renesas/r8a7795-cpg-mssr.c
	drivers/gpu/drm/sti/sti_hqvdp.c
	drivers/media/platform/vsp1/vsp1_dl.c
	drivers/media/usb/au0828/au0828-core.c
	drivers/sh/pm_runtime.c
	drivers/soc/Kconfig
	drivers/soc/Makefile
	drivers/soc/renesas/Kconfig
	drivers/soc/renesas/Makefile
	drivers/soc/renesas/pm-rcar.c
	include/linux/soc/renesas/pm-rcar.h
	include/linux/soc/renesas/rcar-sysc.h
	sound/usb/card.c
	sound/usb/media.c

+++ include/media/media-device.h
@@ -357,7 +357,7 @@ struct media_device {
 	u32 hw_revision;
 	u32 driver_version;
 
-	u32 topology_version;
+	u64 topology_version;
 
 	u32 id;
 	struct ida entity_internal_idx;
@@ -494,7 +494,7 @@ int __must_check __media_device_register(struct media_device *mdev,
 #define media_device_register(mdev) __media_device_register(mdev, THIS_MODULE)
 
 /**
- * __media_device_unregister() - Unegisters a media device element
+ * media_device_unregister() - Unregisters a media device element
  *
  * @mdev:	pointer to struct &media_device
  *
